FinBERT_LLM/requirements.txt
psycopg2-binary
pandas
torch
transformers
tqdm
python-dotenv
google-generativeai